# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: thelake API
  version: 0.1.0
  description: |
    Authenticated HTTP API exposed by thelake. Interactive UI: `GET /swagger`
    (this document is also served as `GET /openapi.json`).

    This API accepts telemetry payloads on standard OTLP HTTP paths:
    - `POST /v1/traces`
    - `POST /v1/logs`

    It also documents the tenant-scoped schema promotion endpoint:
    - `POST /v1/promotions/apply`

    Evaluations are written through:
    - `POST /v1/scores`
    - `GET|POST /v1/score-configs`

    Query surfaces:
    - `POST /v1/traces/search`
    - `POST /v1/traces/details`
    - `POST /v1/spans/search`
    - `GET /v1/spans/{span_id}`
    - `GET /v1/traces/{trace_id}`
    - `POST /v1/sessions/search`
    - `POST /v1/sessions/details`
    - `POST /v1/sessions/summary/rebuild`
    - `GET /v1/sessions/{session_id}`
    - `GET /v1/sessions/{session_id}/recording`

    Control plane:
    - `POST /v1/workspaces` (admin bearer)
    - `GET /v1/meta`
    - `POST /v1/promotions/apply`
    - `GET /v1/data/ducklake-connection`

    Also inventoried here (summary-level): health/ready, `POST /v1/query/sql`,
    and `/v1/fields`. Loki/Tempo Grafana-compat query routes are served by
    the process but are not part of this product OpenAPI document.

    For JSON payloads, requests are expected to follow the OpenTelemetry collector export
    request shapes (`ExportTraceServiceRequest` and `ExportLogsServiceRequest`).

...

...

...

...
    `docs/how-to/promotion.md`.
...
servers:
...
security:
  - bearerAuth: []
paths:
  /slack/events:
    post:
      tags: [integrations]
      summary: Receive Slack Events API callbacks
      description: |
        Slack verifies this endpoint using the configured signing secret. It
        accepts URL verification challenges and authorized evaluator commands
        from Slack event callbacks. This endpoint does not use bearer auth.
      operationId: receiveSlackEvent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              additionalProperties: true
              description: Slack Events API envelope, signed over the raw body.
      responses:
        "200":
          description: Challenge response or event acknowledged
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Malformed Slack event envelope
        "401":
          description: Missing, stale, or invalid Slack request signature
        "403":
          description: Slack team or author is not authorized
        "503":
          description: Slack integration is unavailable; Slack may retry the event
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...